Move Google Cloud Run worker-ID plugin into a .WorkerId sibling project
Cloud Run hosts multiple plugins, so the worker-ID plugin should not own the root Temporalio.Extensions.Gcp.CloudRun namespace. Move it into a new Temporalio.Extensions.Gcp.CloudRun.WorkerId sibling project mirroring the existing .OpenTelemetry sibling, and delete the now-empty root project.

Public type names are unchanged; only the namespace becomes Temporalio.Extensions.Gcp.CloudRun.WorkerId. The new sibling is wired into docfx.json, the ApiDoc index, and the nuget-package workflow to match how .OpenTelemetry is set up.
