+++ .github/workflows/nuget-package.yml
@@ -164,6 +164,8 @@ jobs:
             src/Temporalio.Extensions.Aws.Lambda.OpenTelemetry/bin/Release/*.snupkg
             src/Temporalio.Extensions.Gcp.CloudRun.OpenTelemetry/bin/Release/*.nupkg
             src/Temporalio.Extensions.Gcp.CloudRun.OpenTelemetry/bin/Release/*.snupkg
+            src/Temporalio.Extensions.Gcp.CloudRun.WorkerId/bin/Release/*.nupkg
+            src/Temporalio.Extensions.Gcp.CloudRun.WorkerId/bin/Release/*.snupkg
             src/Temporalio.Extensions.DiagnosticSource/bin/Release/*.nupkg
             src/Temporalio.Extensions.DiagnosticSource/bin/Release/*.snupkg
             src/Temporalio.Extensions.Hosting/bin/Release/*.nupkg
