+++ tests/Temporalio.Tests/Extensions/Gcp/CloudRun/WorkerId/WorkerIdPluginTests.cs
@@ -1,8 +1,8 @@
-namespace Temporalio.Tests.Extensions.Gcp.CloudRun;
+namespace Temporalio.Tests.Extensions.Gcp.CloudRun.WorkerId;
 
 using Temporalio.Client;
 using Temporalio.Common;
-using Temporalio.Extensions.Gcp.CloudRun;
+using Temporalio.Extensions.Gcp.CloudRun.WorkerId;
 using Temporalio.Tests.Extensions.OpenTelemetry;
 using Temporalio.Worker;
 using Xunit;
