+++ tests/Temporalio.Tests/Temporalio.Tests.csproj
@@ -27,7 +27,7 @@
     <ProjectReference Include="..\..\src\Temporalio.Extensions.DiagnosticSource\Temporalio.Extensions.DiagnosticSource.csproj" />
     <ProjectReference Include="..\..\src\Temporalio.Extensions.Aws.Lambda\Temporalio.Extensions.Aws.Lambda.csproj" />
     <ProjectReference Include="..\..\src\Temporalio.Extensions.Aws.Lambda.OpenTelemetry\Temporalio.Extensions.Aws.Lambda.OpenTelemetry.csproj" />
-    <ProjectReference Include="..\..\src\Temporalio.Extensions.Gcp.CloudRun\Temporalio.Extensions.Gcp.CloudRun.csproj" />
+    <ProjectReference Include="..\..\src\Temporalio.Extensions.Gcp.CloudRun.WorkerId\Temporalio.Extensions.Gcp.CloudRun.WorkerId.csproj" />
     <ProjectReference Include="..\..\src\Temporalio.Extensions.Gcp.CloudRun.OpenTelemetry\Temporalio.Extensions.Gcp.CloudRun.OpenTelemetry.csproj" />
     <ProjectReference Include="..\..\src\Temporalio.Extensions.Hosting\Temporalio.Extensions.Hosting.csproj" />
     <ProjectReference Include="..\..\src\Temporalio.Extensions.OpenTelemetry\Temporalio.Extensions.OpenTelemetry.csproj" />
